[improvements]: Refactoring and removed the duplicate code
